[FEA] Account -> User로 객체 이름 변경
 - hostUser or sendUser 등이 좀더 가독성 좋아보임
