--- src/main/java/com/carrot/auction/domain/account/domain/entity/AccountRole.java
@@ -1,5 +1,0 @@
-package com.carrot.auction.domain.account.domain.entity;
-
-public enum AccountRole {
-    ADMIN, USER
-}
