--- src/main/java/com/carrot/auction/domain/auction/controller/AuctionController.java
@@ -1,12 +1,0 @@
-package com.carrot.auction.domain.auction.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/auction")
-public class AuctionController {
-
-}
