+++ src/main/java/com/carrot/auction/domain/user/domain/entity/UserRole.java
@@ -1,0 +1,8 @@
+package com.carrot.auction.domain.user.domain.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum UserRole {
+    ADMIN, USER
+}
